Far simpler implementation of back word.
This fixes many problems that back word was having when using the
gotoPreviousWord function. Back word will now not get stuck when the
current line starts with whitespace. It will also now not skip over some
of the words that start on the beginning of the line. It will stop going
back when it hits an empty line like in vim. It will also skip over
lines that only have whitespace (just like vim).