Adds "V" (linewise visual mode).
Vibreoffice can now select by lines with "V" (visual_line mode). When in
this mode lines can be selected with the j and k keys.

This feature isn't perfect yet. Lines cannot yet be selected properly
with the arrow keys and empty lines at the end of a selection are not
always selected properly.