Fix for "V" (visual select by line).
Vibreoffice can now correctly select any line with "V" (visual select by
line) using the j and k keys. The only downside to this fix is that when
selecting lines downwards with "j" the cursor will always be on the line
bellow the selection. Not a big deal, just not aesthetically pleasing.