Fix for "a" and "p".
"a" will now not move the cursor right before going into insert mode
when the cursor is at the end of the paragraph. Without this fix "a"
would go to the next paragraph before entering insert mode.

"p" will now not move the cursor right before pasting when the cursor is
at the end of the paragraph. Without this fix "p" would go to the next
paragraph before pasting.