Fix for "e".
VibreOffice defines the cursor at being at the end of a word when the
cursor is to the right of the last character. Pressing "i" after
following this definition would start inserting after the word and not
before the last character like it does in vim. This fix changes the "e"
behavior so it now will move one character to the left after reaching
LibreOffice's definition of end of word in cases when it is appropriate.